cambios setter getter  array

+++ app/Http/Controllers/CajaController.php
@@ -7,8 +7,10 @@ use App\MovePayment;
 use App\MoveClosing; 
 use App\CashBox;
 use App\Sell;
+use App\User;
 
 class CajaController extends Controller{  
+	       
 
  	public function funcion(){
  		$alert="";   
@@ -32,28 +34,46 @@ class CajaController extends Controller{
  	}
  	//eliminar cajas
  	public function deleteCaja($parameters){
-    	$cajas = CashBox::find($parameters);
-    	$cajas->delete();
-    	return view('cajas.allCajas');
+		$cajas= CashBox::find($parameters);
+        if ($cajas->id != null){
+			$cajas->delete();			
+			return redirect()->route('allcaja')->with([
+				'mensaje' => 'Caja eliminada csorrectamente!!',
+				'tipo' => 'info'
+			]);
+        }  else {
+			return redirect()->route("allcaja")->with([
+				'mensaje' => 'No se encontro registro',
+				'tipo' => 'danger'
+			]);
+		}   
     }
     //guardar cambios de modificacion de cajas
     public function saveCangecajas(Request $request){         
     	$id = $request->inputId;
-    	$cajas = CashBox::find($id);
-    	$cajas->descripcion = $request->inputDescripcion;
-    	$cajas->status = $request->stcaja;    	  	 
-    	$cajas->save();  
+		$cajas = CashBox::find($id);
+		if($cajas->id != null){
+			$cajas->descripcion = $request->inputDescripcion;
+    		$cajas->status = $request->stcaja;    	  	 
+    		$cajas->save(); 
+			return redirect()->route("allcaja")->with([
+				'mensaje' => 'Datos actualizados correctamente!!',
+				'tipo' => 'info'
+			]);
+		} else {
+			return redirect()->route("allcaja")->with([
+				'mensaje' => 'Los datos no se han actualizado!!',
+				'tipo' => 'danger'
+			]);			 
+		}
     }
     public function  editCajas($param){
-
     	$res=  CashBox::getBox($param);   
-   		return view('cajas.editCajas',['res'=>$res]);
-   		 
+   		return view('cajas.editCajas',['res'=>$res]);   		 
   	}
  	public function cajasAjax(Request $request){    
 	    if($request->ajax())	    {
 	    	$output = '';
-
 	      	$query = $request->get('query');
 	      	if($query != '')
 	      	{
@@ -97,11 +117,9 @@ class CajaController extends Controller{
  		$validate=$this->validate($request,[     
         	'inputCaja' => 'required|unique:cajas,descripcion',        	 
         	'inputStatus'=>'required|nullable'
-      	]);
- 		 
+      	]); 		 
  		 $nameCaja=$request->inputCaja;
  		 //recoger valor seleccionado  ion
- 		 //$status = $request->input('inputStatus');
  		 $status = $request->get('inputStatus');
  		 $caja = new CashBox;
  		 $caja->descripcion=$nameCaja;
@@ -111,9 +129,8 @@ class CajaController extends Controller{
  	}
 		 
 	public function saveMovimientoscaja($id_caja,$inicial,$id_usu,$fechaHora){
-		$total_venta=$this->totalVenta();
-		//guardar id de usuario  usuario que continuaremos turno en una sesion
-		//declarada          
+		$total_venta=$this->totalVenta();	 
+	    
         session(['id_caja' => $id_caja]); 
         $sesionId_caja= session('id_caja');  
 		if ($total_venta==null){
@@ -123,21 +140,15 @@ class CajaController extends Controller{
 		}	
 		 
 		$suma=$inicial+$total_venta;
-		//Para crear un registro.
+		$status='abierto';
+		//Crear registro.
 		$movimientos = new MovePayment;		 
-		$movimientos->id_caja=$id_caja;
-		$movimientos->id_usu=$id_usu;		 
-		$movimientos->dinero_inicial=$inicial;
-		$movimientos->acomulado_ventas= 0;
-		$movimientos->acomulado_entradas=0;
-		$movimientos->acomulado_salidas=0;
-		$movimientos->efectivo_cierre=0;
-		$movimientos->total_caja=0;
-		$movimientos->numero_ventas=0;
-		$movimientos->status='abierto';
-		$movimientos->inicio_en=$fechaHora;
-		$movimientos->termino_en=0; 
-		$movimientos->save();
+		$movimientos->setIdCajaAttribute($id_caja);
+		$movimientos->setIdUsuAttribute($id_usu);		 
+		$movimientos->setDineroInicialAttribute($inicial);
+		$movimientos->setStatusAttribute($status);
+		$movimientos->setInicioEnAttribute($fechaHora);		 
+		$movimientos->save(); 
 	}
 
     
@@ -149,7 +160,7 @@ class CajaController extends Controller{
 	//sacar el total de las ventas del ultimo registro
 	public function totalVenta(){
 		$venta = Sell::latest('id')->first();
-        $total_venta=$venta["total"];       
+        $total_venta=$venta['total'];       
         return  $total_venta;
 	}
 	// Obtener usuario logueado y fechaHora
@@ -158,7 +169,7 @@ class CajaController extends Controller{
 		$id_user=Auth()->user()->id_employee;          
         //fecha y hora
         $time = time();
-	    $fechaHora=date("d-m-Y H:i:s", $time);
+	    $fechaHora=date('d-m-Y H:i:s', $time);
         //almacenar en arreglo los datos a enviar
         $datos=[$id_user,$fechaHora];
         return $datos;
@@ -166,47 +177,51 @@ class CajaController extends Controller{
 	
 	public function altaRegistroFin(Request $request){ 		 
  		$efectivoCaja= $request->efectivoFinal;	
- 		$sesionUserTurno = session('id_usuTurno'); 
+ 		$id_user = session('id_usuTurno'); 
 		$datos=$this->obtenerDatos();
 
 		//sacar datos del arreglo
 		$fechaHora=$datos[1]; 
 		//consulta para comparar total en caja con efectivo al cierre
-		$efectivo=MovePayment::getTurnoOpen($sesionUserTurno);
+		$efectivo=MovePayment::getTurnoOpen($id_user);
 
 		$totalSuma=$efectivo[0]->efectivo_cierre;
-		$id_Mov=$efectivo[0]->id;
+		$id_mov=$efectivo[0]->id;
 		//contar total de id, si hay mas de 3 bloquear pantalla y solicitar contraseña al admin		 
-	 	$ids=MoveClosing::getTotalId($id_Mov);
+	 	$ids=MoveClosing::getTotalId($id_mov);
 		
 		//compara efectivo fisico en caja con el total en la bd
-		if($totalSuma==$efectivoCaja){
-
-			$consultaMovcaja = MovePayment::getTurnoOpen($sesionUserTurno); 	
-			foreach ($consultaMovcaja as $key => $value) {}
-			$idCaja=$value->id_caja; 
-        	$caja = CashBox::updateBoxActive($idCaja);
-			//obtiene registro del  movimiento con usuario logueado y status abierto y modifica	 
-			$Movcaja= MovePayment::updateTurnoOpen($sesionUserTurno, $efectivoCaja, $fechaHora);
-			//destruir sesion
-			session()->forget('id_usuTurno'); 
-			Auth::logout();
-      		return redirect('/'); 
+		if($totalSuma == $efectivoCaja){
+			//obtener id de caja con turno abierto
+			$consultaMovcaja = MovePayment::getTurnoOpen($id_user); 			 
+			if($consultaMovcaja != null){				 
+				$idCaja=$consultaMovcaja[0]->id_caja;
+				$caja = CashBox::updateBoxActive($idCaja);
+				//obtiene registro del  movimiento con usuario logueado y status abierto y modifica	 
+				$Movcaja= MovePayment::updateTurnoOpen($id_user, $efectivoCaja, $fechaHora);
+				//destruir sesion
+				session()->forget('id_usuTurno'); 
+				Auth::logout();
+      			return redirect('/'); 
+			}			 
 
 		} elseif( $ids>=3){
 			// si hay 3 registros con el mismo id regresar ventana para solicitar al admin el desbloqueo
 			return view('cajas.registroFinal')->with('alert', 'Mensaje no Vacio'); 
 
-		}elseif($ids<=3){			 
-			//insertar en la bd los movimientos de cierre
-			$movimientos = new MoveClosing;		 
-			$movimientos->id_user=$sesionUserTurno;
-			$movimientos->id_mov=$id_Mov;			 
-	 		$movimientos->fechaHora=$fechaHora; 
-			$movimientos->save();
+		}elseif($ids<=3){	
+			$data=[
+				'id_user'=>$id_user,
+				'id_mov'=>$id_mov,
+				'fechaHora'=>$fechaHora];
+				var_dump($data);
+				 
+			MoveClosing::create($data);
+			return view('cajas.registroFinal')->with('alert'); 
+			 		 
 		}
 
- 	} //fin metodo
+ 	} 
 
     public function envia(Request $request){
     	//recibir id del usuario que continuaremos turno
@@ -214,9 +229,7 @@ class CajaController extends Controller{
 		$sesionId_usuTurno = session('id_usuTurno');   	 
 		//guardar id de usuario  usuario que continuaremos turno en una sesion
     	session(['id_usuTurno' => $id_usuTurno]);   
-
-    	//
-    	if ($request->input("cerrarTurno") == "closeTurno"){  
+    	if ($request->input('cerrarTurno') == 'closeTurno'){  
     		 
     		//usar funcion para enviar variable alert
     		 $this->funcion();
@@ -229,21 +242,9 @@ class CajaController extends Controller{
     	}
 
     }
-    public function mantenerTurno(){
-		$movi = MovePayment::latest('id')->first();	 			 
-		$movi->acomulado_ventas= 0;
-		$movi->acomulado_entradas=0;
-		$movi->acomulado_salidas=0;
-		$movi->efectivo_cierre=0;
-		$movi->total_caja=0;
-		$movi->numero_ventas=1;
-		$movi->save();
-	}
-
+ 
     //cerrar caja  despues de abrir turno
 	public function cerrarCaja($fechaHora, $sesionId_caja, $id_usu){
-		
-	 	
 		//obtiene registro del  movimiento con usuario logueado y status abierto  
 	 	$Movcaja=MovePayment::updateAll($id_usu,$fechaHora);
 		//actualizar caja a disponible despues de cerrar	
@@ -251,17 +252,14 @@ class CajaController extends Controller{
 		  	 
 	}
 	//funcion Cerrar turno que estuvo abierto
-	public function cerrarTurno($fechaHora,$sesionId_usuTurno){
-		
+	public function cerrarTurno($fechaHora,$sesionId_usuTurno){		
 	 	//obtener datos con sttaus abierto	  		
  	    $consultaMovcaja = MovePayment::getTurnoOpen($sesionId_usuTurno); 
 		foreach ($consultaMovcaja as $key => $value) {}
 		$idCaja=$value->id_caja; 
         $caja = CashBox::updateBoxActive($idCaja);		
-		$Movcaja= MovePayment::updateTurnoOpen($sesionUserTurno, $efectivoCaja=0, $fechaHora);
-	
+		$Movcaja= MovePayment::updateTurnoOpen($sesionUserTurno, $efectivoCaja=0, $fechaHora);	
  	}
-
     	 
 	public function operacionCaja(Request $request){		  
 		$datos=$this->obtenerDatos();
@@ -274,11 +272,11 @@ class CajaController extends Controller{
         $id_caja =$request->caja;       
         //modificar status de caja  ponerla  inactiva
 		$caja = CashBox::updateBoxInactive($id_caja);        
-        if ($request->input("registrar") == "regCaja"){   	      	        	       	  
+        if ($request->input('registrar') == 'regCaja'){   	      	        	       	  
         	$this->saveMovimientoscaja($id_caja,$inicial,$id_usu,$fechaHora);
         	return view('panel.panel');         	
       	} 
-      	if ($request->input("cerrar") == "close"){
+      	if ($request->input('cerrar') == 'close'){
       		 
       		//id de usuario  que continuaremos turno en una sesion
       		$sesionId_usuTurno = session('id_usuTurno'); 
@@ -302,19 +300,15 @@ class CajaController extends Controller{
       			session()->forget('id_usuTurno');      		 
       			Auth::logout();
       			return redirect('/');  
-      		}
-      		 
+      		}      		 
       	}      		 
       	   
-		if ($request->input("mantener") == "open"){
- 	    	$this->mantenerTurno();        	 
+		if ($request->input('mantener') == 'open'){      	 
       		Auth::logout();
-        	return redirect('/');         	 
+        	return redirect('/');          	 
 	    }  
     }
-    
 
-   
 }
